requirements.txt
pyrogram
tgcrypto
requests
aiohttp
PyYAML
